Задача "База данных игроков"
Рефакторинг:
1) Поменял значения Command с byte на int

+++ PlayerDatabase/Command.cs
@@ -1,6 +1,6 @@
 ﻿namespace PlayerDatabase
 {
-    public enum Command : byte
+    public enum Command : int
     {
         AddPlayer = 1,
         DeletePlayer,
